Changed color calculations and storage to be 0 - 15 for performance, noticable fps increase

+++ ASCIIgL/include/ASCIIgL/renderer/Palette.hpp
@@ -6,7 +6,7 @@
 
 // A palette entry: RGB color and its console enum hex mapping
 struct PaletteEntry {
-    glm::vec3 rgb;      // RGB in [0,1] range
+    glm::ivec3 rgb;      // RGB in [0,255] range
     unsigned short  hex;      // Hex value used in COLOR enum
 };
 
@@ -18,7 +18,7 @@ public:
     Palette();
     Palette(std::array<PaletteEntry, COLOR_COUNT> customEntries);
 
-    glm::vec3 GetRGB(unsigned int idx) const;
+    glm::ivec3 GetRGB(unsigned int idx) const;
     unsigned short GetHex(unsigned int idx) const;
     unsigned short GetFgColor(unsigned int idx) const;
     unsigned short GetBgColor(unsigned int idx) const;
